fix: PR feedback: formating, comments, types
