+++ src/types/index.ts
@@ -1,4 +1,3 @@
 export * from './editor';
 export * from './panel';
 export * from './resource';
-export * from './row';
